+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added sensuctl commands to manage the `APIKey` resource.
 - Added support for api keys to be used in api authentication.
 
+## [5.14.2] - 2019-11-04
+
 ### Changed
 - Upgraded etcd to 3.3.17